Clean up the release cleaner

+++ scripts/rel2fw.sh
@@ -33,6 +33,19 @@ if [ ! -e "$FWUP" ]; then
     exit 1
 fi
 
+# If the toolchain contains a pre-built version of mksquashfs,
+# use it; otherwise look for one in the path.
+MKSQUASHFS=$NERVES_TOOLCHAIN/usr/bin/mksquashfs
+[ -e "$MKSQUASHFS" ] || MKSQUASHFS=$(command -v mksquashfs || echo "/usr/bin/mksquashfs")
+if [ ! -e "$MKSQUASHFS" ]; then
+    echo "$SCRIPT_NAME: ERROR: Please install mksquashfs first"
+    echo
+    echo "For example:"
+    echo "   sudo apt-get install squashfs"
+    echo "   brew install squashfs"
+    exit 1
+fi
+
 TMP_DIR=$BASE_DIR/_nerves-tmp
 rm -fr $TMP_DIR
 
@@ -41,16 +54,35 @@ rm -fr $TMP_DIR
 # The RELEASE_DIR can be missing in the case of non-Erlang debug configurations like bbb_linux_defconfig
 [ -z "$FW_FILENAME" ] && FW_FILENAME=${PROJECT_DIR}.fw
 
-# Make sure that the firmware output directories is there.
+# Check that the release directory exists
+if [ ! -d "$RELEASE_DIR" ]; then
+    echo "$SCRIPT_NAME: ERROR: Missing Erlang release directory: ($RELEASE_DIR)"
+    exit 1
+fi
+if [ ! -d "$RELEASE_DIR/lib" -o ! -d "$RELEASE_DIR/releases" ]; then
+    echo "$SCRIPT_NAME: ERROR: Expecting '$RELEASE_DIR' to contain 'lib' and 'releases' subdirectories"
+    exit 1
+fi
+
+# Make sure that the firmware output directories are there.
 mkdir -p `dirname $FW_FILENAME`
 
 # Update the file system bundle
 echo Updating base firmware image with Erlang release...
-$NERVES_ROOT/scripts/create-fs.sh \
-	$NERVES_SDK_IMAGES/rootfs.squashfs \
-	$RELEASE_DIR \
-	$TMP_DIR/rootfs-additions \
-	$TMP_DIR/combined.squashfs
+
+# Create the directory for all of the files that "overlay" the base squashfs
+mkdir -p $TMP_DIR/rootfs-additions
+
+# Construct the proper path for the Erlang/OTP release
+mkdir -p $TMP_DIR/rootfs-additions/srv/erlang
+cp -r $RELEASE_DIR/* $TMP_DIR/rootfs-additions/srv/erlang
+
+# Clean up the Erlang release of all the files that we don't need.
+$NERVES_ROOT/scripts/clean-release.sh $TMP_DIR/rootfs-additions/srv/erlang
+
+# Append the Erlang/OTP release onto the base image.
+cp "$NERVES_SDK_IMAGES/rootfs.squashfs" "$TMP_DIR/combined.squashfs"
+$MKSQUASHFS "$TMP_DIR/rootfs-additions" "$TMP_DIR/combined.squashfs" -no-recovery -no-progress -root-owned >/dev/null
 
 # Build the firmware image
 echo Building $FW_FILENAME...
